Renamed and created new folders for improved project structure
This commit introduced changes to the inventory-microservice and order-management-microservice structure to enhance organization and maintainability. Specifically:

- Renamed existing folders to better reflect their contents
- Created new folders to categorize related files and resources
