+++ inventory-microservice/src/main/java/com/ezzerof/inventorymicroservice/persistance/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.ezzerof.inventorymicroservice.repository;
+package com.ezzerof.inventorymicroservice.persistance.repository;
 
-import com.ezzerof.inventorymicroservice.model.Category;
+import com.ezzerof.inventorymicroservice.persistance.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, String> {
