+++ inventory-microservice/src/main/java/com/ezzerof/inventorymicroservice/persistance/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package com.ezzerof.inventorymicroservice.repository;
+package com.ezzerof.inventorymicroservice.persistance.repository;
 
-import com.ezzerof.inventorymicroservice.model.Product;
+import com.ezzerof.inventorymicroservice.persistance.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
