+++ inventory-microservice/src/main/java/com/ezzerof/inventorymicroservice/service/ProductService.java
@@ -1,6 +1,6 @@
 package com.ezzerof.inventorymicroservice.service;
 
-import com.ezzerof.inventorymicroservice.model.Product;
+import com.ezzerof.inventorymicroservice.persistance.entity.Product;
 
 import java.util.List;
