+++ order-management-microservice/src/main/java/com/ezzerof/ordermanagementmicroservice/persistance/repository/ItemRepository.java
@@ -1,0 +1,7 @@
+package com.ezzerof.ordermanagementmicroservice.persistance.repository;
+
+import com.ezzerof.ordermanagementmicroservice.persistance.entity.Item;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ItemRepository extends JpaRepository<Item, Long> {
+}
