+++ order-management-microservice/src/main/java/com/ezzerof/ordermanagementmicroservice/persistance/repository/OrderRepository.java
@@ -1,0 +1,7 @@
+package com.ezzerof.ordermanagementmicroservice.persistance.repository;
+
+import com.ezzerof.ordermanagementmicroservice.persistance.entity.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+}
