--- order-management-microservice/src/main/java/com/ezzerof/ordermanagementmicroservice/repository/ItemRepository.java
@@ -1,7 +1,0 @@
-package com.ezzerof.ordermanagementmicroservice.repository;
-
-import com.ezzerof.ordermanagementmicroservice.model.Item;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ItemRepository extends JpaRepository<Item, Long> {
-}
