--- order-management-microservice/src/main/java/com/ezzerof/ordermanagementmicroservice/repository/OrderRepository.java
@@ -1,7 +1,0 @@
-package com.ezzerof.ordermanagementmicroservice.repository;
-
-import com.ezzerof.ordermanagementmicroservice.model.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-}
